requirements.txt
numpy>=1.24,<2.0
pandas>=2.0
scikit-learn>=1.3
xgboost>=2.0
flask>=3.0
flask-cors>=4.0
streamlit>=1.30
psycopg2-binary>=2.9
joblib>=1.3
matplotlib>=3.7
seaborn>=0.12
plotly>=5.18
scipy>=1.11
imbalanced-learn>=0.11
python-dotenv>=1.0
requests>=2.31
Pillow>=10.0
